requirements.txt
streamlit
torch
numpy
opencv-python-headless
Pillow
pyyaml
